pyproject.toml
[project]
name = "bookmark-processor"
version = "0.1.0"
dependencies = [
    "typer",
    "prefect",
    "pydantic",
    "requests",
    "beautifulsoup4",
    "readability-lxml",
    "langchain",
    "openai",
    "tiktoken",
    "selenium",
    "webdriver-manager",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "--cov=bookmark_processor --cov-report=term-missing"
pythonpath = ["src"] # Add src to python path for tests to find the package
